[6CORE-101] persistent instand
