+++ 3-platform/warehouse/infra/src/main/java/com/_6core/platform/warehouseinfra/adapter/driven/persistent/entity/WarehouseEntity.java
@@ -1,4 +1,4 @@
-package com._6core.platform.warehouseinfra.adapter.driven.persistence.entity;
+package com._6core.platform.warehouseinfra.adapter.driven.persistent.entity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
