+++ 3-platform/warehouse/infra/src/main/java/com/_6core/platform/warehouseinfra/adapter/driven/persistent/repository/WarehouseReactiveRepository.java
@@ -1,0 +1,6 @@
+package com._6core.platform.warehouseinfra.adapter.driven.persistent.repository;
+
+import com._6core.platform.warehouseinfra.adapter.driven.persistent.entity.WarehouseEntity;
+import org.springframework.data.r2dbc.repository.R2dbcRepository;
+
+public interface WarehouseReactiveRepository extends R2dbcRepository<WarehouseEntity, String> {}
