--- 3-platform/warehouse/infra/src/main/java/com/_6core/platform/warehouseinfra/adapter/persistent/repository/ProductReactiveRepository.java
@@ -1,6 +1,0 @@
-package com._6core.platform.warehouseinfra.adapter.persistent.repository;
-
-import com._6core.platform.warehouseinfra.adapter.driven.persistence.entity.ProductEntity;
-import org.springframework.data.r2dbc.repository.R2dbcRepository;
-
-public interface ProductReactiveRepository extends R2dbcRepository<ProductEntity, String> {}
